+++ WebUI/Areas/Admin/Controllers/CarController.cs
@@ -1,5 +1,6 @@
 ﻿using DtoLayer.BrandDtos;
 using DtoLayer.CarDtos;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Newtonsoft.Json;
@@ -16,10 +17,15 @@ namespace WebUI.Areas.Admin.Controllers
         {
             _httpClientFactory = httpClientFactory;
         }
-
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [Route("Index")]
         public async Task< IActionResult> Index()
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.GetAsync("https://localhost:7250/api/Cars/GetCarsWithBrand");
             if(responseMessage.IsSuccessStatusCode)
@@ -30,11 +36,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
-        
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         [Route("CreateCar")]
         public async Task<IActionResult> CreateCar()
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.GetAsync("https://localhost:7250/api/Brands");
             var jsonData = await responseMessage.Content.ReadAsStringAsync();
@@ -48,10 +59,16 @@ namespace WebUI.Areas.Admin.Controllers
             ViewBag.BrandValues = brandValues;
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [Route("CreateCar")]
         public async Task<IActionResult> CreateCar(CreateCarDto createCarDto)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var jsonData = JsonConvert.SerializeObject(createCarDto);
             var content = new StringContent(jsonData, System.Text.Encoding.UTF8, "application/json");
@@ -62,10 +79,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpDelete]
         [Route("RemoveCar/{id}")]
         public async Task<IActionResult> RemoveCar(int id)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var responseMessage = await client.DeleteAsync($"https://localhost:7250/api/Cars/{id}");
             if(responseMessage.IsSuccessStatusCode)
@@ -74,10 +97,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpGet]
         [Route("UpdateCar/{id}")]
         public async Task<IActionResult> UpdateCar(int id)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
 
             var responseMessage1 = await client.GetAsync("https://localhost:7250/api/Brands");
@@ -100,10 +129,16 @@ namespace WebUI.Areas.Admin.Controllers
             }
             return View();
         }
+        [AllowAnonymous]
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [Route("UpdateCar/{id}")]
         public async Task<IActionResult> UpdateCar(UpdateCarDto updateCarDto)
         {
+            if (!User.Identity.IsAuthenticated || !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Forbidden");
+            }
             var client = _httpClientFactory.CreateClient();
             var jsonData = JsonConvert.SerializeObject(updateCarDto);
             var content = new StringContent(jsonData, System.Text.Encoding.UTF8, "application/json");
